+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix the format of the asset ids: the right format is `str(uuid.uuid4())` and not `uuid.uuid4().hex` (#141)
+
 ## [0.9.0](https://github.com/owkin/connectlib/releases/tag/0.9.0a2) - 2022-04-11
 
 ## Changed